bitmask update -- broken still?
its not finished yet but it looks better, there are however two weird edge case bugs that I just can not figure out. One is when a wall is on the far left it'll see a room neighbor on the next row as being to its direct right?

the other issue is some internal walls wont choose the diagonal tile and go orthogonal instead, but not consistently.